+++ Sources/Widgets/lavabutton.cpp
@@ -2,7 +2,7 @@
 #include "../themehandler.h"
 #include <QtWidgets>
 
-LavaButton::LavaButton(QWidget *parent, double min, double max) : QLabel(parent)
+LavaButton::LavaButton(QWidget *parent, float min, float max) : QLabel(parent)
 {
     this->value = this->min = min;
     this->max = max;
@@ -35,10 +35,10 @@ void LavaButton::setValue(int totalMana, int numCards, int drawCards, int toYour
     if(numCards == 0)    return;
     value = (
                 totalMana +
-                this->drawCards * (totalMana/(float)numCards) +
+                this->drawCards * (totalMana/static_cast<float>(numCards)) +
                 this->toYourHandCards * 4 +
                 this->discoverCards * 4
-             )/(float)numCards;
+             )/static_cast<float>(numCards);
 
     this->value_0_1 = (value - min)/(max - min);
     if(value_0_1 > 1) value_0_1 = 1;
@@ -68,15 +68,16 @@ void LavaButton::paintEvent(QPaintEvent *event)
     QRect targetAll(0, 0, width(), height());
     painter.drawPixmap(targetAll, QPixmap(ThemeHandler::speedCloseFile()));
 
-    QRegion r(QRect(width()*0.15, height()*0.15, width()*0.7, height()*0.7), QRegion::Ellipse);
+    QRegion r(QRect(static_cast<int>(width()*0.15), static_cast<int>(height()*0.15),
+                    static_cast<int>(width()*0.7), static_cast<int>(height()*0.7)), QRegion::Ellipse);
     painter.setClipRegion(r);
     painter.setClipping(true);
-        QRect targetLava(0, (1-value_0_1)*(height()*80/128), width(), height());
+        QRect targetLava(0, static_cast<int>((1-value_0_1)*(height()*80/128)), width(), height());
         painter.drawPixmap(targetLava, QPixmap(ThemeHandler::speedLavaFile()));
     painter.setClipping(false);
 
     QFont font(LG_FONT);
-    font.setPixelSize(width()/3.0);
+    font.setPixelSize(static_cast<int>(width()/3.0));
 
     QPen pen(BLACK);
     pen.setWidth(font.pixelSize()/20);
